feat[#131]: refactoring done
