+++ src/main/java/com/itaxi/server/post/application/dto/PostGetResDto.java
@@ -1,7 +1,6 @@
 package com.itaxi.server.post.application.dto;
 
 import com.itaxi.server.place.application.dto.PlaceResponse;
-import com.itaxi.server.place.domain.Place;
 import com.itaxi.server.post.domain.Joiner;
 import com.itaxi.server.post.domain.Post;
 import com.itaxi.server.post.domain.Stopover;
