+++ src/main/java/com/itaxi/server/post/application/dto/PostGetStopoverInfo.java
@@ -1,6 +1,5 @@
 package com.itaxi.server.post.application.dto;
 
-import com.itaxi.server.post.domain.Joiner;
 import com.itaxi.server.post.domain.Stopover;
 import lombok.Getter;
