+++ src/main/java/com/itaxi/server/post/domain/repository/PostRepository.java
@@ -1,13 +1,11 @@
 package com.itaxi.server.post.domain.repository;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import com.itaxi.server.post.domain.Post;
 
 import com.itaxi.server.place.domain.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
     //List<Person> findByLastname(String lastname);
